edit success and index

+++ view.php
@@ -7,8 +7,8 @@
 
      // Get attendee by Id
      if(!isset($_GET['id'])){
-        //include 'includes/errormessage.php';
-        echo "<h1 class='text-danger'>Please check details and try again</h1>";
+        include 'includes/errormessage.php';
+       // echo "<h1 class='text-danger'>Please check details and try again</h1>";
         
      } else{
         $id = $_GET['id'];
@@ -16,15 +16,14 @@
        
      
 ?>
-<img src="<?php echo empty($result['avatar_path']) ? "uploads/blank.png" : $result['avatar_path']; ?>" 
-class="rounded-circle" style="width: 20%; height: 20%" />
+<img src="<?php echo empty($result['avatar_path']) ? "uploads/blank.png" : $result['avatar_path'] ; ?>" class="rounded-circle" style="width: 20%; height: 20%" />
 
 <div class="card" style="width: 18rem;">
     <div class="card-body">
         <h5 class="card-title">
             <?php echo $result['firstname'] . ' ' . $result['lastname'];  ?>
         </h5>
-        <h6 class="card-subtitle mb-3 text-muted">
+        <h6 class="card-subtitle mb-2 text-muted">
             <?php echo $result['name'];  ?>    
         </h6>
         <p class="card-text">
@@ -39,18 +38,12 @@ class="rounded-circle" style="width: 20%; height: 20%" />
 
     </div>
 </div>
-
 <br/>
         <a href="viewrecords.php" class="btn btn-info">Back to List</a>
-        <a href="edit.php?id=<?php //echo $result['attendee_id'] ?>" class="btn btn-warning">Edit</a>
-        <a onclick="return confirm('Are you sure you want to delete this record?');
-        " href="delete.php?id=<?php //echo $result['attendee_id'] ?>" class="btn btn-danger">Delete</a>
-
-    
-<br>
-<?php } ?>
+        <a href="edit.php?id=<?php echo $result['attendee_id'] ?>" class="btn btn-warning">Edit</a>
+        <a onclick="return confirm('Are you sure you want to delete this record?');" href="delete.php?id=<?php echo $result['attendee_id'] ?>" class="btn btn-danger">Delete</a>
+    <?php } ?>
 <br>
 <br>
 <br>
-
 <?php require_once 'includes/footer.php'; ?>
